requirements.txt
pandas
numpy
scikit-learn
matplotlib
psycopg2-binary
